sunrise_alarm/requirements.txt
"""
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-Migrate==4.0.5
APScheduler==3.10.4
RPi.GPIO==0.7.1
Werkzeug==2.3.7
WTForms==3.0.1
"""